consumer/requirements.txt
kafka-python
python-dotenv
numpy
pandas
scikit-learn
tensorflow
